Extract AutocompleteEngine from CoreReplApp.Autocomplete partial
Move the entire autocomplete suggestion engine (1237 lines) into a
standalone AutocompleteEngine class including token classification,
candidate collection, live hints, and input analysis. Move shared
utilities (IsGlobalOptionToken, TokenizeInputSpans, TokenSpan) into
AutocompleteEngine as internal members.
